pyproject.toml
[project]
name = "sql-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain[google-genai]>=1.1.0",
    "langchain-community>=0.4.1",
    "langgraph>=1.0.4",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "fastapi>=0.122.0",
    "uvicorn[standard]>=0.38.0",
    "psycopg2-binary>=2.9.11",
    "fastapi-cache>=0.1.0",
    "langchain-openai>=1.1.0",
    "langchain-deepseek>=1.0.1",
    "langchain-core>=1.1.0",
    "langchainhub>=0.1.21",
    "langchain-classic>=1.0.0",
    "prophet>=1.2.1",
    "rich>=14.2.0",
]
